Move function to plugin
